+++ docs/changelog.md
@@ -4,7 +4,7 @@ _Copyright (c) 2013 Daniele Veneroni. Released under MIT License._
 
 ### Update 24/05/2013
 
-* Add new units: Catapult, Horseman, Galleon, Knight
+* Add new units: Catapult, Horseman, Galleon, Knight, Spearman, Swordsman, Trireme, Lancer
 * Add unit obsolescence
 
 ### Update 23/05/2013